Alexander/refactoring (#47)
* Breaking things out

* feature: much cleaner implementation
