+++ backend/src/routes/posts.ts
@@ -44,18 +44,21 @@ router.post('/', requireAuth, upload.single('image'), async (req, res) => {
     if (caption.length > 2200) {
       return res.status(400).json({ error: 'Caption must be 2200 characters or less' });
     }
+    
     const mediaUrl = await postService.uploadImage(
       req.file.buffer,
       req.file.mimetype,
       req.user!._id!.toString()
     );
 
+    const federationContext = (req as any).federationContext;
+
     const post = await postService.createPost({
       authorId: req.user!._id!.toString(),
       caption,
       mediaUrl,
       mediaType: req.file.mimetype,
-    });
+    }, federationContext);
 
     const populatedPost = await postService.getPostById(post._id.toString());
